Account redesign, nav drawer added, login design
