+++ .idea/assetWizardSettings.xml
@@ -18,7 +18,7 @@
                                 <PersistentState>
                                   <option name="values">
                                     <map>
-                                      <entry key="url" value="jar:file:/home/fritz/android-studio-ide-173.4819257-linux/android-studio/plugins/android/lib/android.jar!/images/material_design_icons/notification/ic_sms_black_24dp.xml" />
+                                      <entry key="url" value="jar:file:/C:/Program%20Files/Android/Android%20Studio/plugins/android/lib/android.jar!/images/material_design_icons/social/ic_person_black_24dp.xml" />
                                     </map>
                                   </option>
                                 </PersistentState>
@@ -28,8 +28,9 @@
                         </option>
                         <option name="values">
                           <map>
-                            <entry key="outputName" value="default_item" />
-                            <entry key="sourceFile" value="$USER_HOME$" />
+                            <entry key="color" value="800080" />
+                            <entry key="outputName" value="ic_person_black_24dp" />
+                            <entry key="sourceFile" value="C:\Users\omenacle_tech" />
                           </map>
                         </option>
                       </PersistentState>
