+++ app/build.gradle
@@ -30,16 +30,18 @@ dependencies {
     implementation 'com.android.support:design:28.0.0'
     implementation 'com.android.support:support-vector-drawable:28.0.0'
     implementation 'com.android.support:support-v4:28.0.0'
+    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
 
     //butter knife
-   // implementation 'com.jakewharton:butterknife:10.1.0'
+    // implementation 'com.jakewharton:butterknife:10.1.0'
     //annotationProcessor 'com.jakewharton:butterknife-compiler:10.1.0'
     implementation 'com.android.volley:volley:1.1.0'
     implementation 'com.android.support:recyclerview-v7:28.0.0'
     implementation 'gun0912.ted:tedpermission:2.2.2'
+    implementation 'de.hdodenhof:circleimageview:2.1.0'
 
 
 }
