+++ app/src/main/java/zuoix/com/zoomed/activities/AboutActivity.java
@@ -1,0 +1,15 @@
+package zuoix.com.zoomed.activities;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+import zuoix.com.zoomed.R;
+
+public class AboutActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_about);
+    }
+}
