+++ app/src/main/java/zuoix/com/zoomed/activities/MainActivity.java
@@ -1,15 +1,25 @@
 package zuoix.com.zoomed.activities;
 
 import android.content.Context;
+import android.content.Intent;
+import android.net.ConnectivityManager;
+import android.net.NetworkInfo;
+import android.net.Uri;
+import android.support.design.widget.NavigationView;
 import android.support.v4.app.Fragment;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.design.widget.BottomNavigationView;
 import android.support.v4.app.FragmentManager;
+import android.support.v4.view.GravityCompat;
+import android.support.v4.widget.DrawerLayout;
+import android.support.v7.app.AlertDialog;
 import android.support.v7.app.AppCompatActivity;
 import android.view.MenuItem;
 import android.support.v7.widget.Toolbar;
 
+
+import java.net.URL;
 
 import zuoix.com.zoomed.GPSTracker;
 import zuoix.com.zoomed.R;
@@ -30,6 +40,8 @@ public class MainActivity extends AppCompatActivity {
     Fragment active = command;
     Context context;
     SharedPref sp;
+    private DrawerLayout mDrawerLayout;
+
 
     private BottomNavigationView.OnNavigationItemSelectedListener mOnNavigationItemSelectedListener
             = new BottomNavigationView.OnNavigationItemSelectedListener() {
@@ -63,11 +75,13 @@ public class MainActivity extends AppCompatActivity {
         toolbar = findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
         sp = new SharedPref(this);
-        getSupportActionBar().setTitle(R.string.app_name);
+        //getSupportActionBar().setTitle(R.string.app_name);
 
         getSupportActionBar().setDisplayShowHomeEnabled(true);
-        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+        //getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+
         navigation = findViewById(R.id.navigation);
+
         navigation.setOnNavigationItemSelectedListener(mOnNavigationItemSelectedListener);
         context = this;
         fm.beginTransaction().add(R.id.main_container, account, "1").hide(account).commit();
@@ -89,8 +103,63 @@ public class MainActivity extends AppCompatActivity {
             }
 
 
+        mDrawerLayout = findViewById(R.id.container);
+        NavigationView mNavigationView = findViewById(R.id.nav_view);
+        mNavigationView.setNavigationItemSelectedListener(new NavigationView.OnNavigationItemSelectedListener() {
+            @Override
+            public boolean onNavigationItemSelected(@NonNull MenuItem menuItem) {
+                switch (menuItem.getItemId()) {
+                    case R.id.nav_about:
+                        startActivity(new Intent(MainActivity.this, AboutActivity.class));
+                        menuItem.setChecked(true);
+                        mDrawerLayout.closeDrawers();
+                        break;
+                    case R.id.nav_rate:
+                        // TODO replace the empty quotes with the url of app on playstore
+                        openUrl("play.google.com");
+                        menuItem.setChecked(true);
+                        mDrawerLayout.closeDrawers();
+                        break;
+                    case R.id.nav_suggest:
+                        menuItem.setChecked(true);
+                        mDrawerLayout.closeDrawers();
+                        break;
+                    case R.id.nav_privacy:
+                        openUrl("google.com/#privacy");
+                        menuItem.setChecked(true);
+                        mDrawerLayout.closeDrawers();
+                        break;
+
+                }
+                menuItem.setChecked(true);
+                mDrawerLayout.closeDrawers();
+                return true;
+            }
+        });
 
     }
+
+
+
+    private void openUrl(String url){
+        if(isNetworkAvailable()){
+            Intent browserIntent = new Intent(Intent.ACTION_VIEW, Uri.parse(url));
+            startActivity(browserIntent);
+        }
+        else{
+            AlertDialog.Builder bd = new AlertDialog.Builder(MainActivity.this);
+            bd.setMessage(getString(R.string.require_internet));
+            bd.create().show();
+        }
+    }
+
+    private boolean isNetworkAvailable() {
+        ConnectivityManager connectivityManager
+                = (ConnectivityManager) getSystemService(Context.CONNECTIVITY_SERVICE);
+        NetworkInfo activeNetworkInfo = connectivityManager.getActiveNetworkInfo();
+        return activeNetworkInfo != null && activeNetworkInfo.isConnected();
+    }
+
     @Override
     protected void onResume() {
         super.onResume();
@@ -99,12 +168,16 @@ public class MainActivity extends AppCompatActivity {
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
-        if(item.getItemId() == android.R.id.home){
+        /*if(item.getItemId() == android.R.id.home){
             if(active != command){
                 navigation.setSelectedItemId(R.id.command);
             }else{
                 finish();
             }
+        }*/
+        if (item.getItemId() == android.R.id.home) {
+            mDrawerLayout.openDrawer(GravityCompat.START);
+            //return true;
         }
         return super.onOptionsItemSelected(item);
     }
