+++ app/src/main/java/zuoix/com/zoomed/activities/WelcomeActivity.java
@@ -1,12 +1,19 @@
 package zuoix.com.zoomed.activities;
 
+import android.Manifest;
 import android.annotation.SuppressLint;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.graphics.Color;
+import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
+import android.provider.Settings;
+import android.support.annotation.NonNull;
+import android.support.v4.app.ActivityCompat;
+import android.support.v4.content.ContextCompat;
 import android.support.v4.view.PagerAdapter;
 import android.support.v4.view.ViewPager;
 import android.support.v7.app.AlertDialog;
@@ -27,7 +34,9 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import zuoix.com.zoomed.R;
 
@@ -39,6 +48,17 @@ public class WelcomeActivity extends AppCompatActivity {
     private LinearLayout dotsLayout;
     private Button btnNext, btnSkip;
 
+    //ist of all permissions for the app
+    String[] appPermissions = {
+            Manifest.permission.ACCESS_NETWORK_STATE,
+            Manifest.permission.ACCESS_FINE_LOCATION,
+            Manifest.permission.CALL_PHONE,
+            Manifest.permission.INTERNET,
+            Manifest.permission.SEND_SMS,
+            Manifest.permission.READ_PHONE_STATE
+    };
+
+    private static final int PERMISSION_REQUEST_CODE = 1240;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -58,85 +78,192 @@ public class WelcomeActivity extends AppCompatActivity {
 
             setContentView(R.layout.activity_welcome);
 
-            prefManager.setDefaultSim(-1);
-            if (prefManager.getDefaultSim() == -1) {
-                final SmsManager[] smsManager = {null};
-                List<String> simNames = new ArrayList<>();
-                if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.LOLLIPOP_MR1) {
-                    //if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
-                    SubscriptionManager subscriptionManager = (SubscriptionManager) getSystemService(Context.TELEPHONY_SUBSCRIPTION_SERVICE);
-                    @SuppressLint("MissingPermission") final List<SubscriptionInfo> subscriptionInfoList = subscriptionManager.getActiveSubscriptionInfoList();
-                    for (int i = 1; i < subscriptionInfoList.size() + 1; i++) {
-                        simNames.add("Sim " + i);
+
+            //Check for app permissions
+            //In case one or more permissions are not granted
+            //ActivityCompat.requestPermissions() will request permissions
+            //and the control goes to onRequestPermissionResult() callback method
+
+            if (checkAndRequestPermissions()) {
+                prefManager.setDefaultSim(-1);
+                if (prefManager.getDefaultSim() == -1) {
+                    final SmsManager[] smsManager = {null};
+                    List<String> simNames = new ArrayList<>();
+                    if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.LOLLIPOP_MR1) {
+                        //if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
+                        SubscriptionManager subscriptionManager = (SubscriptionManager) getSystemService(Context.TELEPHONY_SUBSCRIPTION_SERVICE);
+                        @SuppressLint("MissingPermission") final List<SubscriptionInfo> subscriptionInfoList = subscriptionManager.getActiveSubscriptionInfoList();
+                        for (int i = 1; i < subscriptionInfoList.size() + 1; i++) {
+                            simNames.add("Sim " + i);
+                        }
+
+                        AlertDialog.Builder builder = new AlertDialog.Builder(this);
+                        builder.setCancelable(false);
+                        builder.setTitle(R.string.select_sim)
+                                .setItems(simNames.toArray(new String[simNames.size()]), new DialogInterface.OnClickListener() {
+                                    public void onClick(DialogInterface dialog, int pos) {
+                                        // The 'which' argument contains the index position
+                                        // of the selected item
+                                        int subId = subscriptionInfoList.get(pos).getSubscriptionId();// change index to 1 if you want to get Subscrption Id for slot 1.
+                                        smsManager[0] = SmsManager.getSmsManagerForSubscriptionId(subId);
+                                        prefManager.setDefaultSim(smsManager[0].getSubscriptionId());
+                                        Log.d("SmsManager", String.valueOf(smsManager[0].getSubscriptionId()));
+                                    }
+                                });
+                        builder.create();
+                        builder.show();
+
+                    } else {
+                        smsManager[0] = SmsManager.getDefault();
                     }
 
-                    AlertDialog.Builder builder = new AlertDialog.Builder(this);
-                    builder.setCancelable(false);
-                    builder.setTitle(R.string.select_sim)
-                            .setItems(simNames.toArray(new String[simNames.size()]), new DialogInterface.OnClickListener() {
-                                public void onClick(DialogInterface dialog, int pos) {
-                                    // The 'which' argument contains the index position
-                                    // of the selected item
-                                    int subId = subscriptionInfoList.get(pos).getSubscriptionId();// change index to 1 if you want to get Subscrption Id for slot 1.
-                                    smsManager[0] = SmsManager.getSmsManagerForSubscriptionId(subId);
-                                    prefManager.setDefaultSim(smsManager[0].getSubscriptionId());
-                                    Log.d("SmsManager", String.valueOf(smsManager[0].getSubscriptionId()));
-                                }
-                            });
-                    builder.create();
-                    builder.show();
-
-                } else {
-                    smsManager[0] = SmsManager.getDefault();
                 }
 
-            }
+                viewPager = findViewById(R.id.view_pager);
+                dotsLayout = findViewById(R.id.layoutDots);
+                btnNext = findViewById(R.id.btn_next);
+                btnSkip = findViewById(R.id.btn_skip);
 
-            viewPager = findViewById(R.id.view_pager);
-            dotsLayout = findViewById(R.id.layoutDots);
-            btnNext = findViewById(R.id.btn_next);
-            btnSkip = findViewById(R.id.btn_skip);
+                // layouts of all welcome sliders
+                // add few more layouts if you want
+                layouts = new int[]{
+                        R.layout.welcome_slide0,
+                        R.layout.welcome_slide1,
+                        R.layout.welcome_slide3,
+                        R.layout.welcome_slide5,
+                        R.layout.welcome_slide6
+                };
 
-            // layouts of all welcome sliders
-            // add few more layouts if you want
-            layouts = new int[]{
-                    R.layout.welcome_slide0,
-                    R.layout.welcome_slide1,
-                    R.layout.welcome_slide3,
-                    R.layout.welcome_slide5,
-                    R.layout.welcome_slide6
-            };
+                // adding bottom dots
+                addBottomDots(0);
 
-            // adding bottom dots
-            addBottomDots(0);
+                // making notification bar transparent
+                changeStatusBarColor();
 
-            // making notification bar transparent
-            changeStatusBarColor();
+                MyViewPagerAdapter myViewPagerAdapter = new MyViewPagerAdapter();
+                viewPager.setAdapter(myViewPagerAdapter);
+                viewPager.addOnPageChangeListener(viewPagerPageChangeListener);
 
-            MyViewPagerAdapter myViewPagerAdapter = new MyViewPagerAdapter();
-            viewPager.setAdapter(myViewPagerAdapter);
-            viewPager.addOnPageChangeListener(viewPagerPageChangeListener);
+                btnNext.setOnClickListener(new View.OnClickListener() {
+                    @Override
+                    public void onClick(View view) {
 
-            btnNext.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View view) {
-
-                    int current = getItem(+1);
-                    if (current < layouts.length) {
-                        // move to next screen
-                        viewPager.setCurrentItem(current);
-                    } else {
+                        int current = getItem(+1);
+                        if (current < layouts.length) {
+                            // move to next screen
+                            viewPager.setCurrentItem(current);
+                        } else {
+                            launchHomeScreen();
+                        }
+                    }
+                });
+                btnSkip.setOnClickListener(new View.OnClickListener() {
+                    @Override
+                    public void onClick(View view) {
                         launchHomeScreen();
                     }
-                }
-            });
-            btnSkip.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View view) {
-                    launchHomeScreen();
-                }
-            });
+                });
+            }
         }
+    }
+
+    private boolean checkAndRequestPermissions() {
+        //Check which permissions are granted
+        List<String> listPermissionsNeeded = new ArrayList<>();
+        for (String pem : appPermissions){
+            if (ContextCompat.checkSelfPermission(this, pem) != PackageManager.PERMISSION_GRANTED){
+                listPermissionsNeeded.add(pem);
+            }
+        }
+
+        //ask for non-granted permissions
+        if (!listPermissionsNeeded.isEmpty()) {
+            ActivityCompat.requestPermissions(this, listPermissionsNeeded.toArray(new String[listPermissionsNeeded.size()]), PERMISSION_REQUEST_CODE);
+            return false;
+        }
+        return true;
+    }
+
+    @Override
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
+        if (requestCode == PERMISSION_REQUEST_CODE) {
+            HashMap<String, Integer> permissionResults = new HashMap<>();
+            int deniedCount = 0;
+
+            //Gather permission grant result
+            for(int i=0; i<grantResults.length; i++) {
+                //Add only permissions which are denied
+                if(grantResults[i] == PackageManager.PERMISSION_DENIED) {
+                    permissionResults.put(permissions[i], grantResults[i]);
+                    deniedCount++;
+                }
+            }
+
+            //check if all permissions are granted
+            if (deniedCount != 0) {
+                for (Map.Entry<String, Integer> entry : permissionResults.entrySet()) {
+                    String permName = entry.getKey();
+                    int permResult = entry.getValue();
+                    //permission is denied (this is the first time, when "never ask again" is not checked)
+                    //so ask again explaining the usage for the permission
+                    //shouldShowRequestPermissionRationale will return true
+                    if (ActivityCompat.shouldShowRequestPermissionRationale(this, permName)) {
+                        showDialog("", "This app needs Location and Internet permission to get map for offline use, Call and SMS  permission offline to work without problems", "Yes, Grant permissions",
+                                new DialogInterface.OnClickListener() {
+                                    @Override
+                                    public void onClick(DialogInterface dialog, int i) {
+                                        dialog.dismiss();
+                                        checkAndRequestPermissions();
+                                    }
+                                },
+                                "No, Exit app", new DialogInterface.OnClickListener(){
+                                    @Override
+                                    public void onClick(DialogInterface dialog, int i) {
+                                        dialog.dismiss();
+                                        finish();
+                                    }
+                                }, false);
+                    }
+                    //permission is denied (and never ask again is checked)
+                    //shouldShowRequestPermissionRationale will return false
+                    else {
+                        //Ask user to go to setting and manually allow permissions
+                        showDialog("", "You have denied some permissions. Allow all permissions at [Setting] > [Permissions]", "Go to Setting", new DialogInterface.OnClickListener() {
+                                    @Override
+                                    public void onClick(DialogInterface dialog, int which) {
+                                        dialog.dismiss();
+                                        //Go to app settings
+                                        Intent intent = new Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS, Uri.fromParts("package", getPackageName(), null));
+                                        intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+                                        startActivity(intent);
+                                        finish();
+                                    }
+                                },
+                                "No, Exit app", new DialogInterface.OnClickListener() {
+                                    @Override
+                                    public void onClick(DialogInterface dialog, int which) {
+                                        dialog.dismiss();
+                                        finish();
+                                    }
+                                }, false);
+                        break;
+                    }
+                }
+            }
+        }
+    }
+
+    private AlertDialog showDialog(String title, String msg, String positiveLabel, DialogInterface.OnClickListener positiveOnClickListener, String negativeLable, DialogInterface.OnClickListener negativeOnClickListener1, boolean b) {
+        AlertDialog.Builder builder = new AlertDialog.Builder(this);
+        builder.setTitle(title);
+        builder.setCancelable(false);
+        builder.setMessage(msg);
+        builder.setPositiveButton(positiveLabel, positiveOnClickListener);
+        builder.setNegativeButton(negativeLable, negativeOnClickListener1);
+
+        AlertDialog alert = builder.create();
+        alert.show();
+        return alert;
     }
 
     private int getItem(int i) {
@@ -202,7 +329,7 @@ public class WelcomeActivity extends AppCompatActivity {
 
     private void launchHomeScreen() {
         prefManager.setFirstTimeLaunch(false);
-        startActivity(new Intent(WelcomeActivity.this, MainActivity.class));
+        startActivity(new Intent(WelcomeActivity.this, LoginActivity.class));
         finish();
     }
