+++ app/src/main/res/drawable/about_icon.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android" xmlns:aapt="http://schemas.android.com/aapt"
+    android:viewportWidth="612"
+    android:viewportHeight="612"
+    android:width="64dp"
+    android:height="64dp">
+    <path
+        android:pathData="M306 153c-21.114 0 -38.25 17.136 -38.25 38.25S284.886 229.5 306 229.5s38.25 -17.117 38.25 -38.25S327.133 153 306 153zM306 267.75c-21.114 0 -38.25 17.117 -38.25 38.25v114.75c0 21.114 17.136 38.25 38.25 38.25s38.25 -17.136 38.25 -38.25V306C344.25 284.886 327.133 267.75 306 267.75zM306 0C137.279 0 0 137.279 0 306s137.279 306 306 306s306 -137.279 306 -306S474.721 0 306 0zM306 554.625C168.893 554.625 57.375 443.088 57.375 306S168.893 57.375 306 57.375c137.088 0 248.625 111.537 248.625 248.625S443.088 554.625 306 554.625z"
+        android:fillColor="#800080" />
+</vector>
