+++ app/src/main/res/layout/activity_login.xml
@@ -1,83 +1,104 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-    <RelativeLayout
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <ScrollView
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:id="@+id/title"
-            style="@style/Titletext"
-            android:text="Company Name"/>
-        <Button
-            android:id="@+id/cancel"
-            android:layout_width="45dp"
-            android:layout_height="45dp"
-            android:layout_marginRight="10dp"
-            android:layout_marginTop="10dp"
-            android:layout_alignParentRight="true"
-            />
+        android:layout_height="wrap_content"
+        android:layout_centerVertical="true">
+
         <LinearLayout
-            android:layout_below="@+id/title"
             android:layout_width="match_parent"
-            android:layout_centerVertical="true"
             android:layout_height="wrap_content"
-            android:orientation="vertical">
+            android:orientation="vertical"
+            android:paddingHorizontal="20dp"
+            android:paddingVertical="10dp"
+            android:paddingTop="10dp">
+
+            <ImageView
+                android:id="@+id/img_profile"
+                android:layout_width="match_parent"
+                android:layout_height="100dp"
+                android:layout_gravity="center_horizontal|center_vertical"
+                android:layout_marginBottom="20dp"
+                android:elevation="1dp"
+                android:scaleType="fitXY"
+                android:padding="0dp"
+                android:src="@drawable/zoomedlogo" />
+
             <android.support.design.widget.TextInputLayout
                 android:layout_width="match_parent"
-                android:hint="username"
-                style="@style/EditTextStyle"
-                android:layout_height="wrap_content">
+                android:layout_height="wrap_content"
+                android:hint="@string/username">
 
                 <android.support.design.widget.TextInputEditText
                     android:id="@+id/username"
-                    android:textSize="13sp"
                     android:layout_width="match_parent"
-                    android:layout_height="40dp" />
+                    android:layout_height="wrap_content"
+                    android:backgroundTint="@color/colorPrimaryDark"
+                    android:drawableStart="@drawable/ic_person_black_24dp"
+                    android:drawableLeft="@drawable/ic_person_black_24dp"
+                    android:textColorHint="@color/colorPrimaryDark"
+                    android:textSize="18sp" />
             </android.support.design.widget.TextInputLayout>
 
-            <RelativeLayout
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content">
 
-                <android.support.design.widget.TextInputLayout
+            <android.support.design.widget.TextInputLayout
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:backgroundTint="@color/colorPrimaryDark"
+
+                android:hint="@string/password">
+
+                <android.support.design.widget.TextInputEditText
+                    android:id="@+id/password"
                     android:layout_width="match_parent"
-                    android:hint="password"
-                    style="@style/EditTextStyle"
-                    android:layout_height="wrap_content">
-
-                    <android.support.design.widget.TextInputEditText
-                        android:id="@+id/password"
-                        android:textSize="13sp"
-                        android:layout_width="match_parent"
-                        android:layout_height="40dp"
-                        android:inputType="textPassword" />
-                </android.support.design.widget.TextInputLayout>
-                <ImageView
-                    android:id="@+id/visible"
-                    android:layout_centerVertical="true"
-                    android:layout_marginRight="35dp"
-                    android:layout_alignParentRight="true"
-                    android:layout_width="35dp"
-                    android:padding="7dp"
-                    android:layout_height="35dp" />
-            </RelativeLayout>
+                    android:layout_height="wrap_content"
+                    android:drawableStart="@drawable/ic_lock_black_24dp"
+                    android:drawableLeft="@drawable/ic_lock_black_24dp"
+                    android:inputType="numberPassword"
+                    android:textColorHint="@color/colorPrimaryDark"
+                    android:textSize="18sp" />
+            </android.support.design.widget.TextInputLayout>
 
 
             <Button
-                android:id="@+id/login"
-                style="@style/small_border_button"
+                android:id="@+id/loginBtn"
                 android:layout_width="match_parent"
-                android:layout_height="40dp"
-                android:text="Login" />
+                android:layout_height="match_parent"
+                android:layout_marginTop="20dp"
+                android:background="@color/colorPrimaryDark"
+                android:text="@string/log_in"
+                android:textColor="@android:color/white"
+                android:textSize="16sp" />
 
+            <RelativeLayout
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="20dp"
+                android:gravity="center">
+
+                <TextView
+                    android:id="@+id/no_account_text"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:gravity="center"
+                    android:text="@string/don_t_have_an_account"
+                    android:textSize="15sp" />
+
+                <TextView
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:layout_marginLeft="5dp"
+                    android:layout_toEndOf="@id/no_account_text"
+                    android:gravity="center"
+                    android:text="@string/demo_link"
+                    android:textColor="?android:attr/textColorPrimary"
+                    android:textSize="15sp" />
+            </RelativeLayout>
         </LinearLayout>
 
-
-    </RelativeLayout>
-
-
+    </ScrollView>
 </RelativeLayout>
