+++ app/src/main/res/layout/fragment_command.xml
@@ -2,17 +2,60 @@
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:background="@null"
-    android:gravity="center"
-    android:paddingBottom="50sp"
     android:layout_height="match_parent">
-
-
-        <zuoix.com.zoomed.AutoFitGridRecyclerView
+        <!--<LinearLayout
             android:layout_width="match_parent"
-            android:id="@+id/command_list"
-            android:background="@null"
-            android:columnWidth="150dp"
-            android:layout_height="wrap_content"/>
+            android:layout_height="?actionBarSize"
+            android:elevation="@dimen/_10sdp"
+            android:background="@color/colorPrimary"
+            android:weightSum="10"
+            android:id="@+id/commandAppBar"
+            android:orientation="horizontal">
+
+                <ImageView
+                    android:id="@+id/backArrow"
+                    android:layout_width="0dp"
+                    android:layout_height="24dp"
+                    android:layout_weight="1"
+                    android:layout_gravity="center_vertical"
+                    android:layout_marginLeft="10dp"
+                    android:layout_marginRight="10dp"
+                    android:padding="5dp"
+                    android:src="@drawable/back_white" />
+                <TextView
+                    android:layout_width="match_parent"
+                    android:layout_height="match_parent"
+                    android:text="@string/command"
+                    android:textStyle="bold"
+                    android:paddingBottom="4dp"
+                    android:layout_marginLeft="10dp"
+                    android:layout_weight="8"
+                    android:gravity="center_vertical|center_horizontal"
+                    android:textColor="@android:color/white"
+                    android:textSize="18sp" />
+                <ImageView
+                    android:id="@+id/saveButton"
+                    android:layout_width="0dp"
+                    android:layout_height="24dp"
+                    android:layout_weight="1"
+                    android:layout_gravity="center_vertical"
+                    android:layout_marginLeft="10dp"
+                    android:layout_marginRight="10dp"
+                    android:padding="5dp" />
+
+        </LinearLayout>-->
+        <RelativeLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:gravity="center">
+
+                <zuoix.com.zoomed.AutoFitGridRecyclerView
+                    android:layout_width="match_parent"
+                    android:id="@+id/command_list"
+                    android:background="@null"
+                    android:columnWidth="150dp"
+                    android:layout_height="wrap_content"/>
+        </RelativeLayout>
 
 
 </RelativeLayout>
