+++ app/src/main/res/values/colors.xml
@@ -1,27 +1,27 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="colorPrimary">#3F51B5</color>
-    <color name="colorPrimaryDark">#303F9F</color>
+    <color name="colorPrimary">#9932CC</color>
+    <color name="colorPrimaryDark">#800080</color>
     <color name="colorAccent">#FF4081</color>
 
 
 
     <!-- Screens background color-->
-    <color name="bg_screen2">#20d2bb</color>
+    <color name="bg_screen2">#c873f4</color>
     <color name="bg_screen1">#f64c73</color>
     <color name="bg_screen3">#3395ff</color>
     <color name="bg_screen4">#c873f4</color>
-    <color name="bg_screen5">#24beab</color>
+    <color name="bg_screen5">#f64c73</color>
 
 
     <!-- dots inactive colors -->
-    <color name="dot_dark_screen2">#14a895</color>
+    <color name="dot_dark_screen2">#a854d4</color>
     <color name="dot_dark_screen1">#d1395c</color>
     <color name="dot_dark_screen3">#2278d4</color>
     <color name="dot_dark_screen4">#a854d4</color>
 
     <!-- dots active colors -->
-    <color name="dot_light_screen2">#8cf9eb</color>
+    <color name="dot_light_screen2">#e4b5fc</color>
     <color name="dot_light_screen1">#f98da5</color>
     <color name="dot_light_screen3">#93c6fd</color>
     <color name="dot_light_screen4">#e4b5fc</color>
@@ -42,5 +42,4 @@
         <item>@color/dot_dark_screen4</item>
         <item>@color/dot_dark_screen1</item>
     </array>
->>>>>>> 1803d6b406e0aae364110ea80a0d01cdd33b6d0d
 </resources>
