+++ app/src/main/res/values/strings.xml
@@ -7,7 +7,7 @@
 
     <string name="account">Account</string>
     <string name="command">Command</string>
-    <string name="setting">Setting</string>
+    <string name="setting">Settings</string>
     <string name="api_key">AIzaSyDaJC6RemjIEe08j-KkvEPDSpa6ynI9WTQ</string>
 
     <string name="start">GOT IT</string>
@@ -19,4 +19,24 @@
     <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="how_to">Lorem Ipsum</string>
     <string name="welcome_desc">This is lorem ipsum welcom description</string>
+    <string name="my_account">Account</string>
+    <string name="first_name">First name</string>
+    <string name="last_name">Last name</string>
+    <string name="phone_number">Phone number</string>
+    <string name="email">Email</string>
+    <string name="save">Log out</string>
+    <string name="change_password">Change password</string>
+    <string name="about_zoomed">About zoomed</string>
+    <string name="password">Password</string>
+    <string name="account_details">Account details</string>
+    <string name="Location">Location</string>
+    <string name="like">Rate our app</string>
+    <string name="suggestions">Suggestions</string>
+    <string name="require_internet">please check your internet connection</string>
+    <string name="policies">Privacy policy</string>
+    <string name="eula">EULA policy</string>
+    <string name="log_in">Log in</string>
+    <string name="username">username</string>
+    <string name="demo_link">Demo link</string>
+    <string name="don_t_have_an_account">Don\'t have an account?</string>
 </resources>
